+++ src/main/java/dev/blachut/svelte/lang/editor/SvelteAutoPopupHandler.kt
@@ -7,7 +7,7 @@ import com.intellij.codeInsight.lookup.LookupManager
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
-import dev.blachut.svelte.lang.directives.SvelteDirectivesSupport
+import dev.blachut.svelte.lang.directives.SvelteDirectiveSupport
 import dev.blachut.svelte.lang.isSvelteContext
 import dev.blachut.svelte.lang.psi.SvelteHtmlTag
 
@@ -19,7 +19,7 @@ class SvelteAutoPopupHandler : TypedHandlerDelegate() {
         val element = file.findElementAt(editor.caretModel.offset)
         if (element?.parent !is SvelteHtmlTag) return Result.CONTINUE
 
-        if (charTyped == SvelteDirectivesSupport.DIRECTIVE_SEPARATOR || charTyped == SvelteDirectivesSupport.MODIFIER_SEPARATOR) {
+        if (charTyped == SvelteDirectiveSupport.DIRECTIVE_SEPARATOR || charTyped == SvelteDirectiveSupport.MODIFIER_SEPARATOR) {
             AutoPopupController.getInstance(project).scheduleAutoPopup(editor)
             return Result.STOP
         }
